refactor: 외부 api관련 패키지 구조 변경
