+++ src/main/java/roomescape/reservation/external/toss/TossPaymentApiClient.java
@@ -2,21 +2,22 @@ package roomescape.reservation.external.toss;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import roomescape.global.api.AuthToken;
 import roomescape.global.api.CustomRequestUri;
 import roomescape.global.api.CustomRestClient;
-import roomescape.global.api.TossApiErrorResponse;
-import roomescape.global.api.TossAuthToken;
+import roomescape.reservation.external.toss.dto.PaymentConfirmRequest;
+import roomescape.reservation.external.toss.dto.TossApiErrorResponse;
 import roomescape.reservation.domain.PaymentInfo;
-import roomescape.reservation.service.PaymentService;
+import roomescape.reservation.service.PaymentApiClient;
 
 @Service
-public class TossPaymentService implements PaymentService {
+public class TossPaymentApiClient implements PaymentApiClient {
 
     private final String tossSecretKey;
     private final String confirmUri;
     private final CustomRestClient customRestClient;
 
-    public TossPaymentService(
+    public TossPaymentApiClient(
             @Value("${api.toss.secret-key}") String tossSecretKey,
             @Value("${api.toss.payment.uri.confirm}") String confirmUri,
             final CustomRestClient customRestClient
@@ -29,7 +30,7 @@ public class TossPaymentService implements PaymentService {
     @Override
     public PaymentInfo paymentReservation(final PaymentConfirmRequest request) {
         CustomRequestUri customRequestUri = new CustomRequestUri(confirmUri);
-        TossAuthToken tossAuthToken = new TossAuthToken(tossSecretKey);
+        AuthToken tossAuthToken = new TossAuthToken(tossSecretKey);
         return customRestClient.post(
                 customRequestUri,
                 tossAuthToken,
