+++ src/test/java/roomescape/common/GlobalConfig.java
@@ -4,8 +4,8 @@ import java.time.Clock;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Primary;
-import roomescape.fake.FakePaymentService;
-import roomescape.reservation.service.PaymentService;
+import roomescape.fake.FakePaymentApiClient;
+import roomescape.reservation.service.PaymentApiClient;
 
 @TestConfiguration
 public class GlobalConfig {
@@ -18,7 +18,7 @@ public class GlobalConfig {
 
     @Bean
     @Primary
-    public PaymentService testPaymentService() {
-        return new FakePaymentService();
+    public PaymentApiClient testPaymentService() {
+        return new FakePaymentApiClient();
     }
 }
