+++ src/test/java/roomescape/unit/domain/global/TossAuthTokenTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.global.api.TossAuthToken;
+import roomescape.reservation.external.toss.TossAuthToken;
 
 class TossAuthTokenTest {
